--- .github/linters/.ecrc
@@ -1,7 +1,0 @@
-{
-  "Exclude" : ["expected"],
-  "Disable": {
-    "IndentSize": true,
-    "Indentation": true
-  }
-}
