+++ .gitignore
@@ -4,3 +4,4 @@
 .mypy_cache/
 __pycache__
 env
+dist
