+++ cgt_calc/exceptions.py
