+++ cgt_calc/misc.py
