+++ cgt_calc/resources/GBP_USD_monthly_history.csv
