+++ cgt_calc/resources/initial_prices.csv
