+++ cgt_calc/resources/template.tex.j2
