+++ pyproject.toml
@@ -1,0 +1,80 @@
+[tool.poetry]
+name = "cgt-calc"
+version = "0.1.0"
+description = "UK capital gains tax calculator for Schwab and Trading212 accounts"
+authors = ["Ruslan Sayfutdinov <ruslan@sayfutdinov.com>"]
+license = "MIT"
+readme = "README.md"
+homepage = "https://github.com/KapJI/capital_gains_calculator"
+repository = "https://github.com/KapJI/capital_gains_calculator"
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3.8",
+    "Topic :: Office/Business :: Financial",
+    "Typing :: Typed",
+]
+keywords = [
+    "capital gain",
+    "schwab",
+    "trading212",
+]
+
+[tool.poetry.urls]
+"Bug Tracker" = "https://github.com/KapJI/capital_gains_calculator/issues"
+
+[tool.poetry.dependencies]
+python = "^3.8"
+Jinja2 = "^2.11.3"
+
+[tool.poetry.dev-dependencies]
+black = "^20.8b1"
+codespell = "^2.0.0"
+flake8 = "^3.9.1"
+flake8-bugbear = "^21.4.3"
+flake8-use-fstring = "^1.1"
+isort = "^5.8.0"
+mypy = "^0.812"
+pre-commit = "^2.12.1"
+pylint = "^2.7.4"
+pytest = "^6.2.3"
+
+[tool.poetry.scripts]
+cgt-calc = 'cgt_calc.main:main'
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.isort]
+profile = "black"
+force_sort_within_sections = true
+combine_as_imports = true
+
+[tool.pylint.format]
+max-line-length = 88
+
+[tool.pylint.messages_control]
+# Reasons disabled:
+# too-many-* - not enforced for the sake of readability
+# too-few-* - same as too-many-*
+# others should be fixed
+disable = [
+    "too-few-public-methods",
+    "too-many-arguments",
+    "too-many-instance-attributes",
+    "too-many-locals",
+    "too-many-return-statements",
+    "too-many-statements",
+    "too-many-branches",
+    "unnecessary-comprehension",
+    "missing-class-docstring",
+    "missing-module-docstring",
+    "missing-function-docstring",
+    "no-else-return",
+    "line-too-long",
+    "invalid-name",
+    "no-self-use",
+    "fixme",
+    "chained-comparison",
+]
