--- requirements.txt
@@ -1,2 +1,0 @@
-Jinja2==2.11.3
-MarkupSafe==1.1.1
