+++ setup.cfg
@@ -1,0 +1,23 @@
+[flake8]
+exclude = .venv,.git,.tox,docs,venv,bin,lib,deps,build
+doctests = True
+# To work with Black
+max-line-length = 88
+# E501: line too long
+# W503: Line break occurred before a binary operator
+# E203: Whitespace before ':'
+# D202 No blank lines allowed after function docstring
+# W504 line break after binary operator
+ignore =
+    E501,
+    W503,
+    E203,
+    D202,
+    W504
+
+[mypy]
+python_version = 3.8
+
+[codespell]
+quiet-level = 2
+skip = poetry.lock,./.git,./.mypy_cache
